pyproject.toml
[tool.black]
line-length = 79
target-version = ['py310', 'py311', 'py312']

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["voidfindertk", "voidfindertk.*"]
namespaces = false

[project]
name = "voidfindertk"
description = "A package to compare void finders"
version = "0.0.1"
readme = "README.md"
authors = [
    { name = "Federico Bustillos", email = "jorgefedericobustillosrava@gmail.com" },
    { name = "Sebastian Gualpa", email = "srgualpa@unc.edu.ar" },
]
license = { file = "LICENSE.txt" }
keywords = ["void finder", "astronomy", "large scale structure"]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Science/Research",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    # science
    "numpy",
    "pandas",
    "scipy",

    # oop lib
    "attrs",
    "uttrs",

    # interaction with CLI
    "sh",

    # astronomy tools
    "astropy",
    "grispy",

    # visualization
    "matplotlib",
    "seaborn",

    # io
    "pyyaml",

    # Multiprocessing
    "joblib",

]

urls = { Repository = "https://github.com/FeD7791/voidFinderProject.git" }
